fix: align event offsets with websocket ids

+++ crates/lilium-database/src/entities/event_processor_offsets.rs
@@ -5,7 +5,7 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub processor_id: String,
-    pub last_processed_id: i32,
+    pub last_processed_id: i64,
     pub last_processed_at: Option<DateTimeUtc>,
     pub updated_at: DateTimeUtc,
     pub last_processed_timestamp: Option<DateTimeUtc>,
